nagfor warning fixes - lots!!
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@2388 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/octal_mod.f90
@@ -257,8 +257,8 @@ MODULE octal_mod
     real(double), pointer :: bnu(:,:) => null() ! 
     real, pointer :: molAbundance(:) => null() ! molecular abundances ! only 1D because only deal with one molecule at a time
     real, pointer :: convergence(:) => null() ! convergence
-    integer*2, pointer :: levelconvergence(:,:) => null() ! convergence
-    integer*2, pointer :: nsplit(:) => null() ! nsplit
+    integer, pointer :: levelconvergence(:,:) => null() ! convergence
+    integer, pointer :: nsplit(:) => null() ! nsplit
     real(double), pointer, DIMENSION(:) :: NHI => null()           ! neutral H
     real(double),  pointer, DIMENSION(:) :: NHII  => null()           ! HII
     real(double),  pointer, DIMENSION(:) :: NHeI => null()            ! HeI
